Update all examples to match
