+++ aries-site/src/examples/templates/wizard/components/CancellationLayer.js
@@ -2,7 +2,7 @@ import React, { useContext } from 'react';
 import PropTypes from 'prop-types';
 import { Box, Button, Heading, Layer, Text } from 'grommet';
 import { WizardContext } from '.';
-import { defaultFormValues } from '../WizardDemo';
+import { defaultFormValues } from '../WizardExample';
 
 export const CancellationLayer = ({ onSetOpen }) => {
   const { setFormValues } = useContext(WizardContext);
