+++ aries-site/src/examples/templates/wizard/components/StepHeader.js
@@ -1,13 +1,13 @@
 import React, { useContext } from 'react';
 import { Box, Heading, Paragraph, Text } from 'grommet';
 import { WizardContext } from '.';
-import { steps } from '../WizardDemo';
+import { steps } from '../WizardExample';
 
 export const StepHeader = () => {
-  const { activeIndex, activeStep } = useContext(WizardContext);
+  const { activeIndex, activeStep, id } = useContext(WizardContext);
   return (
-    <Box id="ccs-demo" gap="xsmall" flex={false}>
-      <Text color="border" weight="bold">
+    <Box id={id} gap="xsmall" flex={false}>
+      <Text>
         Step {activeStep} of {steps.length}
       </Text>
       <Heading color="text-strong" margin="none" size="small">
