+++ aries-site/src/examples/templates/wizard/index.js
@@ -1,3 +1,3 @@
 export * from './TwoColumnWizardExample';
 export * from './WizardValidationExample';
-export * from './WizardDemo';
+export * from './WizardExample';
